mnt(gallery): move the gallery harness under docs/sphinxext
The prediction gallery is documentation tooling, not shipped library
API, so relocate the package from src/nifreeze/_gallery to
docs/sphinxext/gallery (imported as `gallery`). It is put on PYTHONPATH
via conf.py (docs build), the pytest `pythonpath` (tests), mypy_path
(typecheck), and tools/run_gallery.py (the gallery job); it is no
longer installed in the wheel. Drops the now-moot coverage omit.

X-Refs: #458
