+++ docs/sphinxext/gallery/datasets.py
