+++ docs/sphinxext/gallery/manifest.py
