+++ docs/sphinxext/gallery/registry.py
