+++ docs/sphinxext/gallery/render.py
